Move access key logic out of base RH
